requirements.dev.txt
aiofiles==22.1.0
aiosqlite==0.17.0
appdirs==1.4.4
asttokens==2.1.0
backcall==0.2.0
bcrypt==4.0.1
click==8.1.3
decorator==5.1.1
executing==1.2.0
httptools==0.5.0
ipython==8.6.0
iso8601==1.1.0
jedi==0.18.1
matplotlib-inline==0.1.6
multidict==6.0.2
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
ptpython==3.0.20
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
pypika-tortoise==0.1.6
python-decouple==3.6
pytz==2022.6
PyYAML==6.0
sanic==22.9.1
sanic-ext==22.9.1
sanic-routing==22.8.0
six==1.16.0
stack-data==0.6.0
tortoise-cli==0.1.1
tortoise-orm==0.19.2
traitlets==5.5.0
typing_extensions==4.4.0
ujson==5.5.0
uvloop==0.17.0
wcwidth==0.2.5
websockets==10.4
